# apps/edge-service/openapi/edge-openapi.json
{
  "components": {
    "schemas": {
...
      "BarcodeResult": {
        "additionalProperties": false,
        "description": "Barcode read result.",
        "properties": {
          "status": {
            "enum": [
              "READ",
              "NOT_READ",
              "CONFLICT",
              "NOT_REQUIRED"
            ],
            "title": "Status",
            "type": "string"
          },
          "symbology": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbology"
          },
          "value": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "status"
        ],
        "title": "BarcodeResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InspectionDecision": {
        "additionalProperties": false,
        "description": "Deterministic inspection decision produced by the rule engine.",
        "properties": {
          "business_result": {
            "$ref": "#/components/schemas/BusinessResult"
          },
          "decided_at": {
            "format": "date-time",
            "title": "Decided At",
            "type": "string"
          },
          "internal_decision": {
            "$ref": "#/components/schemas/InternalDecision"
          },
          "low_confidence_components": {
            "items": {
              "type": "string"
            },
            "title": "Low Confidence Components",
            "type": "array"
          },
          "missing_components": {
            "items": {
              "type": "string"
            },
            "title": "Missing Components",
            "type": "array"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "title": "Reason Codes",
            "type": "array"
          }
        },
        "required": [
          "internal_decision",
          "business_result",
          "decided_at"
        ],
        "title": "InspectionDecision",
        "type": "object"
      },
      "InspectionImages": {
        "description": "Image slot URLs plus per-slot lifecycle state (F14).\n\nA PURGED slot carries no content URL so the UI renders an explicit purged\nstate instead of a broken image; UNAVAILABLE covers missing or failed media.",
        "properties": {
          "annotated": {
            "title": "Annotated",
            "type": "string"
          },
          "annotated_status": {
            "default": "UNAVAILABLE",
            "enum": [
              "AVAILABLE",
              "PURGED",
              "UNAVAILABLE"
            ],
            "title": "Annotated Status",
            "type": "string"
          },
          "detection": {
            "title": "Detection",
            "type": "string"
          },
          "detection_status": {
            "default": "UNAVAILABLE",
            "enum": [
              "AVAILABLE",
              "PURGED",
              "UNAVAILABLE"
            ],
            "title": "Detection Status",
            "type": "string"
          },
          "inspection_id": {
            "title": "Inspection Id",
            "type": "string"
          },
          "original": {
            "title": "Original",
            "type": "string"
          },
          "original_status": {
            "default": "UNAVAILABLE",
            "enum": [
              "AVAILABLE",
              "PURGED",
              "UNAVAILABLE"
            ],
            "title": "Original Status",
            "type": "string"
          }
        },
        "required": [
          "inspection_id",
          "original",
          "detection",
          "annotated"
        ],
        "title": "InspectionImages",
        "type": "object"
      },
...
      "InspectionRecord": {
        "additionalProperties": false,
        "description": "Persisted product-level inspection result.",
        "properties": {
          "aggregation_policy_version": {
            "title": "Aggregation Policy Version",
            "type": "string"
          },
          "application_version": {
            "title": "Application Version",
            "type": "string"
          },
          "barcode_result": {
            "$ref": "#/components/schemas/BarcodeResult"
          },
          "completed_at": {
            "format": "date-time",
            "title": "Completed At",
            "type": "string"
          },
          "component_model_checksum_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Component Model Checksum Sha256",
            "type": "string"
          },
          "component_model_version_id": {
            "format": "uuid",
            "title": "Component Model Version Id",
            "type": "string"
          },
          "decision": {
            "$ref": "#/components/schemas/InspectionDecision"
          },
          "device_id": {
            "format": "uuid",
            "title": "Device Id",
            "type": "string"
          },
          "device_sequence": {
            "exclusiveMinimum": 0.0,
            "title": "Device Sequence",
            "type": "integer"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/AggregatedComponentEvidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "frame_quality_summary": {
            "$ref": "#/components/schemas/FrameQualitySummary"
          },
          "inference_metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InferenceMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "inspection_id": {
            "format": "uuid",
            "title": "Inspection Id",
            "type": "string"
          },
          "lifecycle_status": {
            "$ref": "#/components/schemas/InspectionLifecycle"
          },
          "media": {
            "items": {
              "$ref": "#/components/schemas/MediaMetadata"
            },
            "title": "Media",
            "type": "array"
          },
          "processing_ms": {
            "minimum": 0.0,
            "title": "Processing Ms",
            "type": "integer"
          },
          "product_detection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProductDetection"
              },
              {
                "type": "null"
              }
            ]
          },
          "product_model_checksum_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Product Model Checksum Sha256",
            "type": "string"
          },
          "product_model_version_id": {
            "format": "uuid",
            "title": "Product Model Version Id",
            "type": "string"
          },
          "product_resolution": {
            "$ref": "#/components/schemas/ProductResolution"
          },
          "roi_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ROIResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "rule_version_id": {
            "format": "uuid",
            "title": "Rule Version Id",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "synchronization_status": {
            "enum": [
              "LOCAL_ONLY",
              "QUEUED",
              "PARTIAL",
              "SYNCED",
              "FAILED"
            ],
            "title": "Synchronization Status",
            "type": "string"
          }
        },
        "required": [
          "inspection_id",
          "device_id",
          "device_sequence",
          "lifecycle_status",
          "started_at",
          "completed_at",
          "barcode_result",
          "product_resolution",
          "frame_quality_summary",
          "application_version",
          "product_model_version_id",
          "product_model_checksum_sha256",
          "component_model_version_id",
          "component_model_checksum_sha256",
          "rule_version_id",
          "aggregation_policy_version",
          "evidence",
          "decision",
          "synchronization_status",
          "processing_ms"
        ],
        "title": "InspectionRecord",
        "type": "object"
      },
...
      "InspectionSummary": {
        "properties": {
          "barcode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Barcode"
          },
          "business_result": {
            "title": "Business Result",
            "type": "string"
          },
          "completed_at": {
            "title": "Completed At",
            "type": "string"
          },
          "inspection_id": {
            "title": "Inspection Id",
            "type": "string"
          },
          "internal_decision": {
            "title": "Internal Decision",
            "type": "string"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "model_rule_versions": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Model Rule Versions",
            "type": "object"
          },
          "product_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Code"
          },
          "reason_summary": {
            "items": {
              "type": "string"
            },
            "title": "Reason Summary",
            "type": "array"
          },
          "sn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sn"
          },
          "upload_state": {
            "title": "Upload State",
            "type": "string"
          }
        },
        "required": [
          "inspection_id",
          "completed_at",
          "business_result",
          "internal_decision",
          "latency_ms",
          "upload_state"
        ],
        "title": "InspectionSummary",
        "type": "object"
      },
...
      "MediaMetadata": {
        "additionalProperties": false,
        "description": "Media object metadata; bytes are never stored in the database.",
        "properties": {
          "checksum_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Checksum Sha256",
            "type": "string"
          },
          "kind": {
            "enum": [
              "KEY_FRAME",
              "ANNOTATED_FRAME",
              "PRODUCT_ROI",
              "NG_CLIP",
              "ROLLING_VIDEO"
            ],
            "title": "Kind",
            "type": "string"
          },
          "lifecycle": {
            "$ref": "#/components/schemas/MediaLifecycle"
          },
          "media_id": {
            "format": "uuid",
            "title": "Media Id",
            "type": "string"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "relative_path": {
            "title": "Relative Path",
            "type": "string"
          },
          "size_bytes": {
            "minimum": 0.0,
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "media_id",
          "kind",
          "lifecycle",
          "relative_path",
          "mime_type",
          "size_bytes",
          "checksum_sha256"
        ],
        "title": "MediaMetadata",
        "type": "object"
      },
      "Page_InspectionSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InspectionSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[InspectionSummary]",
        "type": "object"
      },
...
      "Problem": {
        "additionalProperties": false,
        "description": "RFC 7807 problem response (contract 05 section 6).",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Errors",
            "type": "array"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "StatisticsSummary": {
        "properties": {
          "ng_count": {
            "title": "Ng Count",
            "type": "integer"
          },
          "pass_count": {
            "title": "Pass Count",
            "type": "integer"
          },
          "pass_rate": {
            "title": "Pass Rate",
            "type": "number"
          },
          "total_inspections": {
            "title": "Total Inspections",
            "type": "integer"
          }
        },
        "required": [
          "total_inspections",
          "pass_count",
          "ng_count",
          "pass_rate"
        ],
        "title": "StatisticsSummary",
        "type": "object"
      },
...
      "TraceabilityView": {
        "properties": {
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/TraceabilityAttempt"
            },
            "title": "Attempts",
            "type": "array"
          },
          "final_status": {
            "title": "Final Status",
            "type": "string"
          },
          "sn": {
            "title": "Sn",
            "type": "string"
          }
        },
        "required": [
          "sn",
          "final_status"
        ],
        "title": "TraceabilityView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AssemblyVision Edge API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/inspections": {
      "get": {
        "operationId": "list_inspections_api_v1_inspections_get",
        "parameters": [
          {
            "in": "query",
            "name": "business_result",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Business Result"
            }
          },
          {
            "in": "query",
            "name": "internal_decision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Internal Decision"
            }
          },
          {
            "in": "query",
            "name": "barcode",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Barcode"
            }
          },
          {
            "in": "query",
            "name": "product",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_InspectionSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Inspections",
        "tags": [
          "inspections"
        ]
      }
    },
    "/api/v1/inspections/{inspection_id}": {
      "get": {
        "operationId": "get_inspection_api_v1_inspections__inspection_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "inspection_id",
            "required": true,
            "schema": {
              "title": "Inspection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Inspection not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Inspection",
        "tags": [
          "inspections"
        ]
      }
    },
    "/api/v1/inspections/{inspection_id}/images": {
      "get": {
        "operationId": "inspection_images_api_v1_inspections__inspection_id__images_get",
        "parameters": [
          {
            "in": "path",
            "name": "inspection_id",
            "required": true,
            "schema": {
              "title": "Inspection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionImages"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Inspection Images",
        "tags": [
          "derived"
        ]
      }
    },
    "/api/v1/inspections/{inspection_id}/media": {
      "get": {
        "operationId": "list_inspection_media_api_v1_inspections__inspection_id__media_get",
        "parameters": [
          {
            "in": "path",
            "name": "inspection_id",
            "required": true,
            "schema": {
              "title": "Inspection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MediaMetadata"
                  },
                  "title": "Response List Inspection Media Api V1 Inspections  Inspection Id  Media Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Inspection Media",
        "tags": [
          "inspections"
        ]
      }
    },
...
    "/api/v1/statistics": {
      "get": {
        "operationId": "statistics_api_v1_statistics_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "line",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Line"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatisticsSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Statistics",
        "tags": [
          "derived"
        ]
      }
    },
    "/api/v1/traceability/{sn}": {
      "get": {
        "operationId": "traceability_api_v1_traceability__sn__get",
        "parameters": [
          {
            "in": "path",
            "name": "sn",
            "required": true,
            "schema": {
              "title": "Sn",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceabilityView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Traceability",
        "tags": [
          "derived"
        ]
      }
    },
...
  }
}